src/swarm_controller/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
azure-ai-projects==2.0.0b3
azure-identity==1.23.0
azure-storage-blob==12.25.1
azure-storage-queue==12.13.0
openai==1.82.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-instrumentation-fastapi==0.54b1
azure-monitor-opentelemetry-exporter==1.0.0b35
